♻️ separate code representation layer tests into multiple specialized files
