--- tests/test_code.py
@@ -1,1108 +1,0 @@
-# Copyright (c) 2023 - 2026 Chair for Design Automation, TUM
-# All rights reserved.
-#
-# SPDX-License-Identifier: MIT
-#
-# Licensed under the MIT License
-
-"""Test the CSSCode class."""
-
-from __future__ import annotations
-
-from typing import TYPE_CHECKING
-
-import numpy as np
-import pytest
-import stim
-from numpy.testing import assert_array_equal
-
-from mqt.qecc import CSSCode, StabilizerCode
-from mqt.qecc.codes import (
-    ConcatenatedCSSCode,
-    InvalidCSSCodeError,
-    InvalidStabilizerCodeError,
-    construct_bb_code,
-    construct_iceberg_code,
-    construct_many_hypercube_code,
-    construct_quantum_hamming_code,
-)
-from mqt.qecc.codes.constructions.rotated_surface_code import InvalidDistanceError, RotatedSurfaceCode
-from mqt.qecc.codes.core.pauli import (
-    InvalidPauliError,
-    Pauli,
-    PauliTableau,
-    StabilizerTableau,
-    complete_stabilizer_tableau_with_destabilizers,
-)
-from mqt.qecc.codes.core.symplectic import (
-    SymplecticMatrix,
-    SymplecticVector,
-    symplectic_product,
-)
-from mqt.qecc.mod2 import are_in_same_coset, is_in_row_space
-
-if TYPE_CHECKING:  # pragma: no cover
-    import numpy.typing as npt
-
-
-def test_pauli() -> None:
-    """Test the Pauli class."""
-    p1 = Pauli.from_pauli_string("XIZ")
-    p2 = Pauli(SymplecticVector(np.array([1, 0, 0, 0, 0, 1])))
-    assert p1 == p2
-    p3 = p1 * p2
-    assert p3 == Pauli.from_pauli_string("III")
-    p4 = Pauli.from_pauli_string("-X")
-    p5 = Pauli.from_pauli_string("+Z")
-    p6 = Pauli.from_pauli_string("Y")
-    assert p4 * p5 != p6
-    assert p4 * p5 == -p6
-
-    assert np.array_equal(p1.x_part(), np.array([1, 0, 0]))
-    assert np.array_equal(p1.z_part(), np.array([0, 0, 1]))
-    assert np.array_equal(p6.x_part(), np.array([1]))
-    assert np.array_equal(p6.z_part(), np.array([1]))
-    assert len(p1) == 3
-    assert len(p6) == 1
-
-    assert p4.anticommute(p5)
-    p7 = Pauli.from_pauli_string("XI")
-    p8 = Pauli.from_pauli_string("IZ")
-    assert p8.commute(p7)
-
-    with pytest.raises(IndexError):
-        p1[3]
-
-
-def test_symplectic() -> None:
-    """Test the SymplecticMatrix and SymplecticVector classes."""
-    ones = SymplecticVector.ones(3)
-    zeros = SymplecticVector.zeros(3)
-    assert ones - ones == zeros
-    assert ones + ones == zeros
-
-    v = SymplecticVector(np.array([1, 0, 0, 0, 0, 1]))
-    w = SymplecticVector(np.array([0, 1, 0, 0, 0, 1]))
-    assert w + v == v + w
-    assert w - v == -v + w
-
-    obj = "abc"
-    assert v != obj
-
-    assert v @ w == 0
-    u = SymplecticVector(np.array([0, 0, 1, 0, 0, 0]))
-    assert v @ u == 1
-
-    eye = SymplecticMatrix.symplectic_identity(3)
-    zero_mat = SymplecticMatrix.zeros(6, 3)
-    assert eye + eye == zero_mat
-    assert eye - eye == zero_mat
-
-    vs = [v.vector, w.vector, u.vector, ones.vector, zeros.vector, v.vector]
-    m = SymplecticMatrix(np.array(vs))
-    assert eye @ m.transpose() == m
-    assert m @ eye == m
-
-    for i, row in enumerate(m):
-        assert np.array_equal(row, vs[i])
-
-    assert m != obj
-    assert len(m) == 6
-    assert m.shape == (6, 6)
-    assert m.n == 3
-
-
-@pytest.mark.parametrize(
-    ("tableau_matrix", "expected"),
-    [
-        (np.array([[1, 0, 0, 0], [0, 1, 0, 0]], dtype=np.int8), True),
-        (np.array([[1, 0, 0, 1], [0, 1, 1, 0]], dtype=np.int8), False),
-    ],
-)
-def test_is_css(tableau_matrix: npt.NDArray[np.int8], expected: bool) -> None:
-    """Test the is_css method."""
-    tableau = StabilizerTableau(SymplecticMatrix(tableau_matrix))
-    assert tableau.is_css() == expected
-
-
-@pytest.mark.parametrize(
-    ("tableau_fixture", "expected_x_part", "expected_z_part"),
-    [
-        ("identity_tableau", np.array([[1], [0]], dtype=np.int8), np.array([[0], [1]], dtype=np.int8)),
-        ("hadamard_tableau", np.array([[0], [1]], dtype=np.int8), np.array([[1], [0]], dtype=np.int8)),
-    ],
-)
-def test_get_x_and_z_parts(
-    tableau_fixture: str,
-    expected_x_part: npt.NDArray[np.int8],
-    expected_z_part: npt.NDArray[np.int8],
-    request: pytest.FixtureRequest,
-) -> None:
-    """Test the get_x_part and get_z_part methods with various tableaus."""
-    tableau = request.getfixturevalue(tableau_fixture)
-    x_part = tableau.get_x_part()
-    z_part = tableau.get_z_part()
-    assert_array_equal(x_part, expected_x_part)
-    assert_array_equal(z_part, expected_z_part)
-
-
-def test_stabilizer_tableau() -> None:
-    """Test the StabilizerTableau class."""
-    with pytest.raises(InvalidPauliError):
-        StabilizerTableau.from_pauli_strings([])
-
-    with pytest.raises(InvalidPauliError):
-        StabilizerTableau.from_paulis([])
-
-    m = SymplecticMatrix(np.array([[1, 0], [0, 1]]))
-    with pytest.raises(InvalidPauliError):
-        StabilizerTableau(m, np.array([1]))
-
-    p1 = Pauli.from_pauli_string("XIZ")
-    p2 = Pauli.from_pauli_string("ZIX")
-    p3 = Pauli.from_pauli_string("IZX")
-    t1 = StabilizerTableau.from_paulis([p1, p2, p3])
-    t2 = StabilizerTableau(np.array([[1, 0, 0, 0, 0, 1], [0, 0, 1, 1, 0, 0], [0, 0, 1, 0, 1, 0]]), np.array([0, 0, 0]))
-    assert t1 == t2
-
-    t3 = StabilizerTableau.from_pauli_strings(["ZII", "IZI", "IIZ"])
-    assert t1 != t3
-
-    t4 = StabilizerTableau.from_pauli_strings(["ZII"])
-    assert t1 != t4
-
-    assert len(t1) == 3
-
-
-@pytest.fixture
-def identity_tableau() -> StabilizerTableau:
-    """Fixture for the identity stabilizer tableau."""
-    return StabilizerTableau.from_matrix(np.eye(2, dtype=np.int8))
-
-
-@pytest.fixture
-def hadamard_tableau() -> StabilizerTableau:
-    """Fixture for the Hadamard stabilizer tableau."""
-    return StabilizerTableau.from_matrix(np.array([[0, 1], [1, 0]], dtype=np.int8))
-
-
-@pytest.fixture
-def cnot_tableau() -> StabilizerTableau:
-    """Fixture for the CNOT stabilizer tableau."""
-    return StabilizerTableau.from_matrix(
-        np.array(
-            [
-                [1, 1, 0, 0],
-                [0, 1, 0, 0],
-                [0, 0, 1, 0],
-                [0, 0, 1, 1],
-            ],
-            dtype=np.int8,
-        )
-    )
-
-
-@pytest.mark.parametrize(
-    ("stim_tableau", "expected_tableau_fixture"),
-    [
-        (stim.Tableau.from_named_gate("I"), "identity_tableau"),
-        (stim.Tableau.from_named_gate("H"), "hadamard_tableau"),
-        (stim.Tableau.from_named_gate("CX"), "cnot_tableau"),
-    ],
-)
-def test_stabilizer_tableau_from_stim_tableau(
-    stim_tableau: stim.Tableau, expected_tableau_fixture: str, request: pytest.FixtureRequest
-) -> None:
-    """Test the from_stim_tableau method of StabilizerTableau."""
-    expected_tableau = request.getfixturevalue(expected_tableau_fixture)
-    tableau = StabilizerTableau.from_stim_tableau(stim_tableau)
-    assert tableau == expected_tableau
-
-
-@pytest.mark.parametrize(
-    ("stim_circuit", "expected_tableau_fixture"),
-    [
-        (stim.Circuit("I 0"), "identity_tableau"),
-        (stim.Circuit("H 0"), "hadamard_tableau"),
-        (stim.Circuit("CX 0 1"), "cnot_tableau"),
-    ],
-)
-def test_stabilizer_tableau_from_stim_circuit(
-    stim_circuit: stim.Circuit, expected_tableau_fixture: str, request: pytest.FixtureRequest
-) -> None:
-    """Test the from_stim_circuit method of StabilizerTableau."""
-    expected_tableau = request.getfixturevalue(expected_tableau_fixture)
-    tableau = StabilizerTableau.from_stim_circuit(stim_circuit)
-    assert tableau == expected_tableau
-
-
-@pytest.fixture
-def rep_code_checks() -> tuple[npt.NDArray[np.int8] | None, npt.NDArray[np.int8] | None]:
-    """Return the parity check matrices for the repetition code."""
-    hx = np.array([[1, 1, 0], [0, 1, 1]])
-    hz = None
-    return hx, hz
-
-
-@pytest.fixture
-def rep_code_checks_reverse() -> tuple[npt.NDArray[np.int8] | None, npt.NDArray[np.int8] | None]:
-    """Return the parity check matrices for the repetition code."""
-    hz = np.array([[1, 1, 0], [0, 0, 1]])
-    hx = None
-    return hx, hz
-
-
-@pytest.fixture
-def steane_code_checks() -> tuple[npt.NDArray[np.int8], npt.NDArray[np.int8]]:
-    """Return the check matrices for the Steane code."""
-    hx = np.array([[1, 1, 1, 1, 0, 0, 0], [1, 0, 1, 0, 1, 0, 1], [0, 1, 1, 0, 1, 1, 0]])
-    hz = hx
-    return hx, hz
-
-
-@pytest.fixture
-def steane_code() -> CSSCode:
-    """Return the Steane code."""
-    hx = np.array([[1, 1, 1, 1, 0, 0, 0], [1, 0, 1, 0, 1, 0, 1], [0, 1, 1, 0, 1, 1, 0]])
-    hz = hx
-    return CSSCode(distance=3, Hx=hx, Hz=hz)
-
-
-@pytest.fixture
-def five_qubit_code_stabs() -> list[str]:
-    """Return the five qubit code."""
-    return ["XZZXI", "IXZZX", "XIXZZ", "ZXIXZ"]
-
-
-@pytest.fixture
-def five_qubit_code() -> StabilizerCode:
-    """Return the five qubit code."""
-    return StabilizerCode(["XZZXI", "IXZZX", "XIXZZ", "ZXIXZ"], 3, z_logicals=["ZZZZZ"], x_logicals=["XXXXX"])
-
-
-def test_invalid_css_codes() -> None:
-    """Test that an invalid CSS code raises an error."""
-    hx = np.array([[1, 1, 1]])
-    hz = np.array([[1, 0, 0]])
-    with pytest.raises(InvalidCSSCodeError):
-        CSSCode(distance=3, Hx=hx, Hz=hz)
-
-    hz = np.array([[1, 1, 0]])
-    with pytest.raises(InvalidCSSCodeError):
-        CSSCode(distance=3, Hx=hx, Hz=hz, x_distance=4, z_distance=1)
-
-    hz = np.array([[1, 1]])
-    with pytest.raises(InvalidCSSCodeError):
-        CSSCode(distance=3, Hx=hx, Hz=hz)
-
-    with pytest.raises(InvalidCSSCodeError):
-        CSSCode(distance=3)
-
-
-@pytest.mark.parametrize("checks", ["steane_code_checks", "rep_code_checks", "rep_code_checks_reverse"])
-def test_logicals(checks: str, request: pytest.FixtureRequest) -> None:
-    """Test the logical operators of the CSSCode class."""
-    hx, hz = request.getfixturevalue(checks)
-    code = CSSCode(distance=3, Hx=hx, Hz=hz)
-    assert code.Lx is not None
-    assert code.Lz is not None
-    assert code.Lx.shape[1] == code.Lz.shape[1] == code.n
-    assert code.Lx.shape[0] == code.Lz.shape[0]
-
-    assert code.Lx @ code.Lz.T % 2 != 0
-
-    if code.Hz is not None:
-        assert np.all(code.Lx @ code.Hz.T % 2 == 0)
-    if code.Hx is not None:
-        assert np.all(code.Lz @ code.Hx.T % 2 == 0)
-
-
-def test_errors(steane_code_checks: tuple[npt.NDArray[np.int8], npt.NDArray[np.int8]]) -> None:
-    """Test error detection and symdromes."""
-    hx, hz = steane_code_checks
-    code = CSSCode(distance=3, Hx=hx, Hz=hz)
-    e1 = np.array([1, 0, 0, 0, 0, 0, 0])
-    e2 = np.array([0, 1, 0, 0, 1, 0, 0])
-    e3 = np.array([0, 0, 0, 0, 0, 1, 1])
-    e4 = np.array([0, 1, 1, 1, 0, 0, 0])
-
-    assert np.array_equal(code.get_x_syndrome(e1), code.get_z_syndrome(e2))
-    assert np.array_equal(code.get_x_syndrome(e2), code.get_z_syndrome(e2))
-
-    x_syndrome_1 = code.get_x_syndrome(e1)
-    x_syndrome_2 = code.get_x_syndrome(e2)
-    x_syndrome_3 = code.get_x_syndrome(e3)
-    x_syndrome_4 = code.get_x_syndrome(e4)
-
-    assert np.array_equal(x_syndrome_1, x_syndrome_2)
-    assert not np.array_equal(x_syndrome_1, x_syndrome_3)
-    assert np.array_equal(x_syndrome_1, x_syndrome_4)
-
-    assert code.check_if_logical_x_error((e1 + e2) % 2)
-    assert code.check_if_logical_z_error((e1 + e2) % 2)
-    assert not code.stabilizer_eq_x_error(e1, e2)
-    assert not code.stabilizer_eq_z_error(e1, e2)
-
-    assert not code.check_if_logical_x_error((e1 + e4) % 2)
-    assert not code.check_if_logical_z_error((e1 + e4) % 2)
-    assert code.stabilizer_eq_x_error(e1, e4)
-    assert code.stabilizer_eq_z_error(e1, e4)
-
-
-def test_rep_code(rep_code_checks: tuple[npt.NDArray[np.int8], npt.NDArray[np.int8]]) -> None:
-    """Test utility functions and correctness of the repetition code."""
-    hx, hz = rep_code_checks
-    code = CSSCode(distance=1, Hx=hx, Hz=hz)
-    assert code.n == 3
-    assert code.k == 1
-    assert code.distance == 1
-    assert not code.is_self_dual()
-
-    e1 = np.array([1, 0, 0], dtype=np.int8)
-    e2 = np.array([0, 1, 0], dtype=np.int8)
-    e3 = np.array([0, 0, 1], dtype=np.int8)
-    assert np.array_equal(code.get_x_syndrome(e1), np.array([1, 0]))
-    assert np.array_equal(code.get_x_syndrome(e2), np.array([1, 1]))
-    assert np.array_equal(code.get_x_syndrome(e3), np.array([0, 1]))
-
-    assert code.get_z_syndrome(e1).size == 0
-
-    assert code.check_if_logical_z_error((e1 + e2 + e3) % 2)
-    assert not code.check_if_x_stabilizer((e1 + e2 + e3) % 2)
-    assert code.check_if_x_stabilizer((e1 + e2) % 2)
-    assert not code.check_if_z_stabilizer((e1 + e2 + e3) % 2)
-    assert not code.check_if_z_stabilizer((e1 + e3) % 2)
-
-    assert code.stabilizer_eq_x_error(e1, (e1 + e2 + e3) % 2)
-    assert not code.stabilizer_eq_z_error(e1, (e1 + e2 + e3) % 2)
-    assert code.stabilizer_eq_z_error(e1, e1)
-
-
-def test_steane(steane_code_checks: tuple[npt.NDArray[np.int8], npt.NDArray[np.int8]]) -> None:
-    """Test utility functions and correctness of the Steane code."""
-    hx, hz = steane_code_checks
-    code = CSSCode(distance=3, Hx=hx, Hz=hz)
-    assert code.n == 7
-    assert code.k == 1
-    assert code.distance == 3
-    assert code.is_self_dual()
-
-    x_paulis = code.x_checks_as_pauli_strings()
-    z_paulis = code.z_checks_as_pauli_strings()
-    assert x_paulis is not None
-    assert z_paulis is not None
-    assert len(x_paulis) == len(z_paulis) == 3
-    assert x_paulis == ["XXXXIII", "XIXIXIX", "IXXIXXI"]
-    assert z_paulis == ["ZZZZIII", "ZIZIZIZ", "IZZIZZI"]
-
-    hx_reordered = hx[::-1, :]
-    code_reordered = CSSCode(distance=3, Hx=hx_reordered, Hz=hz)
-    assert code.is_equivalent(code_reordered)
-
-
-@pytest.mark.parametrize("n", [72, 90, 108, 144, 288])
-def test_bb_codes(n: int) -> None:
-    """Test that BB codes are constructed as valid CSS codes."""
-    code = construct_bb_code(n)
-    assert code.n == n
-    assert code.Hx is not None
-    assert code.Hz is not None
-    assert np.all(code.Hx @ code.Hx.T % 2) == 0
-
-
-def test_five_qubit_code(five_qubit_code_stabs: list[str]) -> None:
-    """Test that the five qubit code is constructed as a valid stabilizer code."""
-    z_logicals = ["ZZZZZ"]
-    x_logicals = ["XXXXX"]
-
-    code = StabilizerCode(five_qubit_code_stabs, distance=3, x_logicals=x_logicals, z_logicals=z_logicals)
-    assert code.n == 5
-    assert code.k == 1
-    assert code.distance == 3
-
-    error = "XIIII"
-    syndrome = code.get_syndrome(error)
-    assert np.array_equal(syndrome, np.array([0, 0, 0, 1]))
-
-    stabilizer_eq_error = "IZZXI"
-    assert code.stabilizer_equivalent(error, stabilizer_eq_error)
-
-    different_error = "IZIII"
-    assert not code.stabilizer_equivalent(error, different_error)
-
-    strings = code.stabs_as_pauli_strings()
-    assert strings == five_qubit_code_stabs
-
-
-def test_stabilizer_sign() -> None:
-    """Test that (negative) signs are correctly handled in stabilizer codes."""
-    s = ["-ZZZZ", "-XXXX"]
-    code = StabilizerCode(s)
-    assert code.n == 4
-    assert code.k == 2
-
-    error = "XIII"
-    syndrome = code.get_syndrome(error)
-    assert np.array_equal(syndrome, np.array([1, 0]))
-
-
-def test_trivial_code() -> None:
-    """Test code with no stabilizers."""
-    code = StabilizerCode.get_trivial_code(3)
-    assert code.n == 3
-    assert code.k == 3
-    assert code.x_logicals.to_pauli_list() == [
-        Pauli.from_pauli_string("XII"),
-        Pauli.from_pauli_string("IXI"),
-        Pauli.from_pauli_string("IIX"),
-    ]
-    assert code.z_logicals.to_pauli_list() == [
-        Pauli.from_pauli_string("ZII"),
-        Pauli.from_pauli_string("IZI"),
-        Pauli.from_pauli_string("IIZ"),
-    ]
-    assert code.generators.n_rows == 0
-
-
-def test_negative_distance() -> None:
-    """Test that an error is raised if a negative distance is provided."""
-    with pytest.raises(InvalidStabilizerCodeError):
-        StabilizerCode(["ZZZZ", "XXXX"], distance=-1)
-
-
-def test_different_length_stabilizers() -> None:
-    """Test that an error is raised if stabilizers have different lengths."""
-    with pytest.raises(InvalidPauliError):
-        StabilizerCode(["ZZZZ", "X", "Y"])
-
-
-def test_invalid_pauli_strings() -> None:
-    """Test that invalid Pauli strings raise an error."""
-    with pytest.raises(InvalidPauliError):
-        StabilizerCode(["ABCD", "XIXI", "YIYI"])
-
-
-def test_no_x_logical() -> None:
-    """Test that an error is raised if no X logical is provided when a Z logical is provided."""
-    with pytest.raises(InvalidStabilizerCodeError):
-        StabilizerCode(["ZZZZ", "XXXX"], z_logicals=["XXII"])
-
-
-def test_no_z_logical() -> None:
-    """Test that an error is raised if no Z logical is provided when an X logical is provided."""
-    with pytest.raises(InvalidStabilizerCodeError):
-        StabilizerCode(["ZZZZ", "XXXX"], x_logicals=["ZZII"])
-
-
-def test_not_compatible_logicals() -> None:
-    """Test that an error is raised if no Z logical is provided when an X logical is provided."""
-    with pytest.raises(InvalidStabilizerCodeError):
-        StabilizerCode([], n=2, z_logicals=["ZI", "IZ"], x_logicals=["XX", "ZZ"])
-
-
-def test_logicals_wrong_length() -> None:
-    """Test that an error is raised if the logicals have the wrong length."""
-    with pytest.raises(InvalidStabilizerCodeError):
-        StabilizerCode(["ZZZZ", "XXXX"], x_logicals=["XX"], z_logicals=["IZZI"])
-    with pytest.raises(InvalidStabilizerCodeError):
-        StabilizerCode(["ZZZZ", "XXXX"], x_logicals=["IXXI"], z_logicals=["ZZ"])
-
-
-def test_commuting_logicals() -> None:
-    """Test that an error is raised if the logicals commute."""
-    with pytest.raises(InvalidStabilizerCodeError):
-        StabilizerCode(["ZZZZ", "XXXX"], z_logicals=["ZZII"], x_logicals=["XXII"])
-
-
-def test_anticommuting_logicals() -> None:
-    """Test that an error is raised if the logicals anticommute with the stabilizer generators."""
-    with pytest.raises(InvalidStabilizerCodeError):
-        StabilizerCode(["ZZZZ", "XXXX"], z_logicals=["ZIII"], x_logicals=["IXXI"])
-    with pytest.raises(InvalidStabilizerCodeError):
-        StabilizerCode(["ZZZZ", "XXXX"], z_logicals=["IZZI"], x_logicals=["XIII"])
-
-
-def test_too_many_logicals() -> None:
-    """Test that an error is raised if too many logicals are provided."""
-    with pytest.raises(InvalidStabilizerCodeError):
-        StabilizerCode(["ZZZZ", "XXXX"], z_logicals=["ZZII", "ZZII", "ZZII"], x_logicals=["IXXI"])
-    with pytest.raises(InvalidStabilizerCodeError):
-        StabilizerCode(["ZZZZ", "XXXX"], z_logicals=["IZZI"], x_logicals=["XXII", "XXII", "XXII"])
-
-
-def test_empty_logicals() -> None:
-    """Test that an the case of empty logicals is handled gracefully."""
-    code = StabilizerCode(["XX", "ZZ"], z_logicals=[], x_logicals=[])
-    assert code.k == 0
-
-
-def test_code_equality() -> None:
-    """Test equality of stabilizer codes."""
-    # Equal codes
-    code1 = StabilizerCode(
-        ["XZZXI", "IXZZX", "XIXZZ", "ZXIXZ"],
-        z_logicals=["ZZZZZ"],
-        x_logicals=["XXXXX"],
-    )
-    code2 = StabilizerCode(
-        ["XZZXI", "IXZZX", "XIXZZ", "ZXIXZ"],
-        z_logicals=["ZZZZZ"],
-        x_logicals=["XXXXX"],
-    )
-    assert code1 == code2  # literally the same code
-
-    # different basis of stabilizers, logicals automatically computed
-    code3 = StabilizerCode(
-        ["XZZXI", "IXZZX", "XIXZZ", "YXXYI"],
-        z_logicals=["ZIXXI"],
-        x_logicals=["XXXXX"],
-    )
-
-    assert code1.is_equivalent(code3)
-
-    # Unequal codes (swapped logical operators)
-    code4 = StabilizerCode(
-        ["XZZXI", "IXZZX", "XIXZZ", "ZXIXZ"],
-        z_logicals=["XXXXX"],
-        x_logicals=["ZZZZZ"],
-    )
-    assert code1 != code4
-
-
-def test_logical_mapping() -> None:
-    """Test the get_logical_mapping method of StabilizerCode."""
-    # Define two equivalent codes
-    code1 = StabilizerCode(
-        ["XZZXI", "IXZZX", "XIXZZ", "ZXIXZ"],
-        z_logicals=["ZZZZZ"],
-        x_logicals=["XXXXX"],
-    )
-    code2 = StabilizerCode(
-        ["XZZXI", "IXZZX", "XIXZZ", "ZXIXZ"],
-        z_logicals=["ZZZZZ"],
-        x_logicals=["XXXXX"],
-    )
-    mapping = code1.get_logical_mapping(code2)
-    assert mapping == [0]
-
-    code3 = StabilizerCode(["XXXX", "ZZZZ"], z_logicals=["ZZII", "IZZI"], x_logicals=["IXXI", "XXII"])
-    code4 = StabilizerCode(["XXXX", "ZZZZ"], z_logicals=["IZZI", "ZZII"], x_logicals=["XXII", "IXXI"])
-
-    mapping = code3.get_logical_mapping(code4)
-    assert mapping == [1, 0]
-
-    code_5 = StabilizerCode(  # first logical is product of logicals
-        ["XXXX", "ZZZZ"], z_logicals=["ZIZI", "IZZI"], x_logicals=["IXXI", "XIXI"]
-    )
-
-    mapping = code3.get_logical_mapping(code_5)
-    assert mapping is None
-
-
-def test_trivial_css_concatenation(steane_code: CSSCode) -> None:
-    """Test that the trivial concatenation of a CSS code is the code itself."""
-    inner_code = CSSCode.get_trivial_code(1)
-    concatenated = ConcatenatedCSSCode(steane_code, inner_code)
-
-    assert concatenated.n == 7
-    assert concatenated.k == 1
-    assert concatenated.distance == 3
-    assert concatenated == steane_code
-
-
-def test_hamming_code() -> None:
-    """Test that the Hamming code is constructed as a valid CSS code."""
-    code = construct_quantum_hamming_code(3)
-    assert code.n == 7
-    assert code.k == 1
-    assert code.distance == 3
-
-
-def test_many_hypercube_code_level_1() -> None:
-    """Test that the many-hypercube code."""
-    code = construct_many_hypercube_code(1)
-    assert code.n == 6
-    assert code.k == 4
-    assert code.distance == 2
-    iceberg = construct_iceberg_code(3)
-    assert code == iceberg
-
-
-def test_many_hypercube_code_level_2() -> None:
-    """Test that the many-hypercube code."""
-    code = construct_many_hypercube_code(2)
-    assert code.n == 36
-    assert code.k == 16
-    assert code.distance == 4
-
-
-def test_stabilizer_code_from_file(tmp_path) -> None:  # type: ignore[no-untyped-def]
-    """Test that a stabilizer code can be constructed from a file."""
-    file_content = "XZZXI\nIXZZX\nXIXZZ\nZXIXZ"
-    file_path = tmp_path / "test_file.txt"
-    file_path.write_text(file_content)
-
-    code = StabilizerCode.from_file(file_path)
-
-    assert code.n == 5
-    assert code.k == 1
-
-
-def test_css_code_from_file(tmp_path) -> None:  # type: ignore[no-untyped-def]
-    """Test that a CSS code can be constructed from a file."""
-    file_content = "XIIXXXI\nIXIIXXX\nIIXXIXX\nZIIZZZI\nIZIIZZZ\nIIZZIZZ"
-    file_path = tmp_path / "test_file.txt"
-    file_path.write_text(file_content)
-
-    code = CSSCode.from_file(file_path)
-
-    assert code.n == 7
-    assert code.k == 1
-
-
-def test_css_code_from_file_empty_line(tmp_path) -> None:  # type: ignore[no-untyped-def]
-    """Test that a CSS code can be constructed from a file."""
-    file_content = "XIIXXXI\nIXIIXXX\n\nIIXXIXX\nZIIZZZI\nIZIIZZZ\nIIZZIZZ"
-    file_path = tmp_path / "test_file.txt"
-    file_path.write_text(file_content)
-
-    code = CSSCode.from_file(file_path)
-
-    assert code.n == 7
-    assert code.k == 1
-
-
-def test_css_code_from_binary_matrix_space_separated(tmp_path) -> None:  # type: ignore[no-untyped-def]
-    """Test loading a CSS code from space-separated binary matrix format."""
-    file_content = """1 1 1 1 0 0 0
-1 0 1 0 1 0 1
-0 1 1 0 1 1 0
-
-1 1 1 1 0 0 0
-1 0 1 0 1 0 1
-0 1 1 0 1 1 0"""
-    file_path = tmp_path / "test_css_binary_space.txt"
-    file_path.write_text(file_content)
-
-    code = CSSCode.from_file(file_path)
-
-    assert code.n == 7
-    assert code.Hx.shape == (3, 7)
-    assert code.Hz.shape == (3, 7)
-    assert np.array_equal(
-        code.Hx, np.array([[1, 1, 1, 1, 0, 0, 0], [1, 0, 1, 0, 1, 0, 1], [0, 1, 1, 0, 1, 1, 0]], dtype=np.int8)
-    )
-    assert np.array_equal(
-        code.Hz, np.array([[1, 1, 1, 1, 0, 0, 0], [1, 0, 1, 0, 1, 0, 1], [0, 1, 1, 0, 1, 1, 0]], dtype=np.int8)
-    )
-
-
-def test_css_code_from_binary_matrix_comma_separated(tmp_path) -> None:  # type: ignore[no-untyped-def]
-    """Test loading a CSS code from comma-separated binary matrix format."""
-    file_content = """1,1,1,1,0,0,0
-1,0,1,0,1,0,1
-0,1,1,0,1,1,0
-
-1,1,1,1,0,0,0
-1,0,1,0,1,0,1
-0,1,1,0,1,1,0
-"""
-    file_path = tmp_path / "test_css_binary_comma.txt"
-    file_path.write_text(file_content)
-
-    code = CSSCode.from_file(file_path)
-
-    assert code.n == 7
-    assert code.Hx.shape == (3, 7)
-    assert code.Hz.shape == (3, 7)
-
-
-def test_css_code_from_binary_matrix_list_notation(tmp_path) -> None:  # type: ignore[no-untyped-def]
-    """Test loading a CSS code from list notation binary matrix format."""
-    file_content = """[[1,1,1,1,0,0,0],
-[1,0,1,0,1,0,1],
-[0,1,1,0,1,1,0]]
-
-[[1,1,1,1,0,0,0],
-[1,0,1,0,1,0,1],
-[0,1,1,0,1,1,0]]"""
-    file_path = tmp_path / "test_css_binary_list.txt"
-    file_path.write_text(file_content)
-
-    code = CSSCode.from_file(file_path)
-
-    assert code.n == 7
-    assert code.Hx.shape == (3, 7)
-    assert code.Hz.shape == (3, 7)
-
-
-def test_css_code_from_binary_matrix_numpy_notation(tmp_path) -> None:  # type: ignore[no-untyped-def]
-    """Test loading a CSS code from numpy array notation format."""
-    file_content = """[[1 1 1 1 0 0 0]
- [1 0 1 0 1 0 1]
- [0 1 1 0 1 1 0]]
-
-[[1 1 1 1 0 0 0]
- [1 0 1 0 1 0 1]
- [0 1 1 0 1 1 0]]"""
-    file_path = tmp_path / "test_css_binary_numpy.txt"
-    file_path.write_text(file_content)
-
-    code = CSSCode.from_file(file_path)
-
-    assert code.n == 7
-    assert code.Hx.shape == (3, 7)
-    assert code.Hz.shape == (3, 7)
-
-
-def test_css_code_from_binary_matrix_x_only(tmp_path) -> None:  # type: ignore[no-untyped-def]
-    """Test loading a CSS code with only X stabilizers from binary format."""
-    file_content = """1,1,1,1,0,0,0
-1,0,1,0,1,0,1
-0,1,1,0,1,1,0"""
-    file_path = tmp_path / "test_css_x_only.txt"
-    file_path.write_text(file_content)
-
-    code = CSSCode.from_file(file_path)
-
-    assert code.n == 7
-    assert code.Hx.shape == (3, 7)
-    assert code.Hz.shape == (0, 7)
-
-
-def test_css_code_from_binary_matrix_empty(tmp_path) -> None:  # type: ignore[no-untyped-def]
-    """Test that an error is raised for empty CSS code files."""
-    file_content = ""
-    file_path = tmp_path / "test_css_empty.txt"
-    file_path.write_text(file_content)
-
-    with pytest.raises(InvalidCSSCodeError, match="File is empty"):
-        CSSCode.from_file(file_path)
-
-
-def test_css_code_from_invalid_pauli_string(tmp_path) -> None:  # type: ignore[no-untyped-def]
-    """Test that an error is raised for invalid Pauli strings in CSS code files."""
-    file_content = "XIIX\nIIII\nZZII"
-    file_path = tmp_path / "test_css_invalid.txt"
-    file_path.write_text(file_content)
-
-    with pytest.raises(InvalidCSSCodeError, match="Invalid stabilizer"):
-        CSSCode.from_file(file_path)
-
-
-def test_stabilizer_code_from_binary_matrix_comma_separated(tmp_path) -> None:  # type: ignore[no-untyped-def]
-    """Test loading a stabilizer code from a comma-separated binary matrix file."""
-    file_content = """1,0,0,1,0,0,1,1,0,0
-0,1,0,0,1,0,0,1,1,0
-1,0,1,0,0,0,0,0,1,1
-0,1,0,1,0,1,0,0,0,1"""
-    file_path = tmp_path / "test_binary_comma.txt"
-    file_path.write_text(file_content)
-
-    code = StabilizerCode.from_file(file_path)
-
-    assert code.n == 5
-    assert code.generators.n_rows == 4
-
-
-def test_stabilizer_code_from_binary_matrix_list_notation(tmp_path) -> None:  # type: ignore[no-untyped-def]
-    """Test loading a stabilizer code from a binary matrix file with list notation."""
-    file_content = """[[1,0,0,1,0,0,1,1,0,0],
-[0,1,0,0,1,0,0,1,1,0],
-[1,0,1,0,0,0,0,0,1,1],
-[0,1,0,1,0,1,0,0,0,1]]"""
-    file_path = tmp_path / "test_binary_list.txt"
-    file_path.write_text(file_content)
-
-    code = StabilizerCode.from_file(file_path)
-
-    assert code.n == 5
-    assert code.generators.n_rows == 4
-
-
-def test_stabilizer_code_from_binary_matrix_correctness(tmp_path) -> None:  # type: ignore[no-untyped-def]
-    """Test that binary matrix loading produces the correct Pauli strings."""
-    file_content = """1 1 1 1"""
-    file_path = tmp_path / "test_binary_correctness.txt"
-    file_path.write_text(file_content)
-
-    code = StabilizerCode.from_file(file_path)
-
-    assert code.n == 2
-    stabs = code.stabs_as_pauli_strings()
-    assert stabs == ["YY"]
-
-
-def test_stabilizer_code_from_binary_matrix_five_qubit(tmp_path) -> None:  # type: ignore[no-untyped-def]
-    """Test loading the five-qubit code from binary matrix format."""
-    file_content = """1 0 0 1 0 0 1 1 0 0
-    0 1 0 0 1 0 0 1 1 0
-    1 0 1 0 0 0 0 0 1 1
-    0 1 0 1 0 1 0 0 0 1"""
-    file_path = tmp_path / "test_five_qubit.txt"
-    file_path.write_text(file_content)
-
-    code = StabilizerCode.from_file(file_path)
-    print(code.stabs_as_pauli_strings())
-
-    assert code.n == 5
-    assert code.generators.n_rows == 4
-
-
-def test_stabilizer_code_from_binary_matrix_invalid_rows(tmp_path) -> None:  # type: ignore[no-untyped-def]
-    """Test that an error is raised for binary matrices with odd number of rows."""
-    file_content = """1 1 0 0
-0 0 1 0"""
-    file_path = tmp_path / "test_binary_invalid.txt"
-    file_path.write_text(file_content)
-
-    with pytest.raises(InvalidStabilizerCodeError, match=r"Stabilizer generators must commute with each other."):
-        StabilizerCode.from_file(file_path)
-
-
-def test_stabilizer_code_from_binary_matrix_empty(tmp_path) -> None:  # type: ignore[no-untyped-def]
-    """Test that an error is raised for empty binary matrix files."""
-    file_content = ""
-    file_path = tmp_path / "test_binary_empty.txt"
-    file_path.write_text(file_content)
-
-    with pytest.raises(InvalidStabilizerCodeError, match="File is empty"):
-        StabilizerCode.from_file(file_path)
-
-
-def test_stabilizer_code_to_tableau(five_qubit_code: StabilizerCode) -> None:
-    """Test the to_tableau method of StabilizerCode."""
-    tableau = five_qubit_code.to_tableau()
-
-    assert tableau.n == 5
-    assert tableau.num_rows() == 6
-
-    x_log_pauli = tableau[0]
-    z_log_pauli = tableau[1]
-    assert not x_log_pauli.commute(z_log_pauli)
-
-    for i in range(2, 6):
-        stab = tableau[i]
-        assert x_log_pauli.commute(stab)
-        assert z_log_pauli.commute(stab)
-
-
-def test_stabilizer_code_to_tableau_trivial() -> None:
-    """Test the to_tableau method for trivial code."""
-    code = StabilizerCode.get_trivial_code(3)
-    tableau = code.to_tableau()
-
-    assert tableau.n == 3
-    assert tableau.num_rows() == 6
-
-    for i in range(3):
-        x_log = tableau[i]
-        z_log = tableau[3 + i]
-        assert not x_log.commute(z_log)
-
-
-@pytest.mark.parametrize(("x_distance", "z_distance"), [(3, 3), (7, 3), (5, 7)])
-def test_rotated_surface_code_params(x_distance: int, z_distance: int) -> None:
-    """Test the RotatedSurfaceCode class."""
-    code = RotatedSurfaceCode(x_distance=x_distance, z_distance=z_distance)
-    n = x_distance * z_distance
-    assert code.n == n
-    assert code.k == 1
-    assert code.x_distance == x_distance
-    assert code.z_distance == z_distance
-
-    num_patches = (x_distance - 1) * (z_distance - 1)
-    num_x_stabs = num_patches // 2 + num_patches % 2 + (z_distance - 1)
-    num_z_stabs = num_patches // 2 + (x_distance - 1)
-    assert code.Hx.shape == (num_x_stabs, n)
-    assert code.Hz.shape == (num_z_stabs, n)
-
-
-def test_rotated_surface_code_invalid_distance() -> None:
-    """Test that an error is raised if the distance is not odd."""
-    with pytest.raises(InvalidDistanceError):
-        RotatedSurfaceCode(distance=6)
-    with pytest.raises(InvalidDistanceError):
-        RotatedSurfaceCode(x_distance=4, z_distance=5)
-    with pytest.raises(InvalidDistanceError):
-        RotatedSurfaceCode(x_distance=5, z_distance=4)
-    with pytest.raises(InvalidDistanceError):
-        RotatedSurfaceCode(x_distance=5, z_distance=None)
-
-
-def test_complete_stabilizer_tableau_with_destabilizers():
-    """Test completing a stabilizer tableau with destabilizers."""
-    stabs = StabilizerTableau.from_pauli_strings(["ZZ"])
-    completed = complete_stabilizer_tableau_with_destabilizers(stabs)
-
-    assert completed.num_rows() == 2
-
-    destab = completed[0]
-    stab = completed[1]
-    assert destab.anticommute(stab)
-
-    stabs = StabilizerTableau.from_pauli_strings(["XXXXIII", "ZZZZIII"])
-    completed = complete_stabilizer_tableau_with_destabilizers(stabs)
-
-    assert completed.num_rows() == 4
-
-    for i in range(2):
-        destab_i = completed[i]
-        for j in range(2):
-            stab_j = completed[2 + j]
-            if i == j:
-                assert destab_i.anticommute(stab_j)
-            else:
-                assert destab_i.commute(stab_j)
-
-    for i in range(2):
-        for j in range(i + 1, 2):
-            assert completed[i].commute(completed[j])
-
-    stabs = StabilizerTableau.from_pauli_strings(["XII", "IZI", "IIX"])
-    completed = complete_stabilizer_tableau_with_destabilizers(stabs)
-
-    assert completed.num_rows() == 6
-
-    for i in range(3):
-        destab_i = completed[i]
-        for j in range(3):
-            stab_j = completed[3 + j]
-            if i == j:
-                assert destab_i.anticommute(stab_j)
-            else:
-                assert destab_i.commute(stab_j)
-
-
-def test_complete_stabilizer_tableau_invalid_cases():
-    """Test error handling for invalid inputs."""
-    stabs = StabilizerTableau.from_pauli_strings(["XX", "ZZ", "YY"])
-
-    with pytest.raises(ValueError, match="Cannot have more stabilizers than qubits"):
-        complete_stabilizer_tableau_with_destabilizers(stabs)
-
-
-def test_compute_logical_steane():
-    """Test that compute_logical correctly finds logicals for Steane code."""
-    stabs = ["XXXXIII", "XIXIXIX", "IXXIXXI", "ZZZZIII", "ZIZIZIZ", "IZZIZZI"]
-    code = StabilizerCode(stabs, distance=3)
-
-    assert code.k == 1
-    assert code.z_logicals is not None
-    assert code.x_logicals is not None
-    assert code.z_logicals.n_rows == 1
-    assert code.x_logicals.n_rows == 1
-
-    assert not code.z_logicals[0].commute(code.x_logicals[0])
-
-
-def test_compute_logical_multiple_pairs():
-    """Test compute_logical with k>1."""
-    stabs = ["XXXX", "ZZZZ"]
-    code = StabilizerCode(stabs, distance=2)
-
-    assert code.k == 2
-    assert code.z_logicals.n_rows == 2
-    assert code.x_logicals.n_rows == 2
-
-
-def test_set_logicals():
-    """Test that set_logicals correctly sets the logical operators."""
-    h = np.array([[1, 1, 1, 1]])
-    code = CSSCode(h, h)
-    assert code.k == 2
-
-    lxs = code.Lx.copy()
-    lzs = code.Lz.copy()
-
-    lxs[0] ^= lxs[1]
-    lzs[1] ^= lzs[0]
-
-    code.set_x_logicals(lxs)
-    code.set_z_logicals(lzs)
-
-    assert np.array_equal(code.Lx, lxs)
-    assert np.array_equal(code.Lz, lzs)
-
-
-def test_symplectic_product_shapes() -> None:
-    """Test symplectic_product for vector and matrix operands."""
-    v = np.array([1, 0, 0, 0, 0, 1], dtype=np.int8)  # XIZ
-    u = np.array([0, 0, 1, 0, 0, 0], dtype=np.int8)  # IIX
-    assert symplectic_product(v, v) == 0
-    assert symplectic_product(v, u) == 1
-
-    m = np.array([v, u], dtype=np.int8)
-    assert_array_equal(symplectic_product(m, v), np.array([0, 1]))
-    assert_array_equal(symplectic_product(v, m), np.array([0, 1]))
-    assert_array_equal(symplectic_product(m, m), np.array([[0, 1], [1, 0]]))
-
-    empty = np.zeros((0, 6), dtype=np.int8)
-    assert symplectic_product(empty, m).shape == (0, 2)
-
-
-def test_symplectic_product_validation() -> None:
-    """Test that symplectic_product rejects invalid inputs."""
-    v = np.array([1, 0, 0, 0, 0, 1], dtype=np.int8)
-    with pytest.raises(ValueError, match="even"):
-        symplectic_product(v, np.array([1, 0, 0], dtype=np.int8))
-    with pytest.raises(ValueError, match="same number of qubits"):
-        symplectic_product(v, np.array([1, 0, 0, 0], dtype=np.int8))
-    with pytest.raises(ValueError, match="dimensions"):
-        symplectic_product(v, np.zeros((1, 2, 6), dtype=np.int8))
-    with pytest.raises(ValueError, match="nonzero"):
-        symplectic_product(np.zeros(0, dtype=np.int8), np.zeros(0, dtype=np.int8))
-
-
-def test_row_space_helpers() -> None:
-    """Test is_in_row_space and are_in_same_coset, including empty bases."""
-    basis = np.array([[1, 1, 0], [0, 1, 1]], dtype=np.int8)
-    assert is_in_row_space(np.array([1, 0, 1], dtype=np.int8), basis)
-    assert not is_in_row_space(np.array([1, 0, 0], dtype=np.int8), basis)
-    assert is_in_row_space(np.zeros(3, dtype=np.int8), basis)
-
-    empty = np.zeros((0, 3), dtype=np.int8)
-    assert is_in_row_space(np.zeros(3, dtype=np.int8), empty)
-    assert not is_in_row_space(np.array([1, 0, 0], dtype=np.int8), empty)
-
-    e1 = np.array([1, 0, 0], dtype=np.int8)
-    e2 = np.array([0, 1, 1], dtype=np.int8)
-    assert are_in_same_coset(e1, (e1 + basis[0]) % 2, basis)
-    assert not are_in_same_coset(e1, e2, empty)
-    assert are_in_same_coset(e1, e1, empty)
-
-
-def test_independent_of_generator_signs() -> None:
-    """Test that the stabilizer group is independent of the signs of the generators.
-
-    Note: Depending on the demand of a sign-aware matrix multiplication, this test can be removed.
-    """
-    plus = StabilizerCode(["XXXX", "ZZZZ"])
-    minus = StabilizerCode(["-XXXX", "-ZZZZ"])
-    assert plus.k == minus.k == 2
-    assert plus.equal_stabilizer_group(minus)
-
-
-def test_pauli_tableau_alias() -> None:
-    """StabilizerTableau remains available as a deprecated alias of PauliTableau."""
-    assert StabilizerTableau is PauliTableau
-    t = PauliTableau.from_pauli_strings(["XX", "ZZ"])
-    assert t.symplectic.shape == (2, 4)
-    assert_array_equal(t.symplectic, t.tableau.matrix)
-
-
-def test_trivial_code_syndromes_and_cosets() -> None:
-    """Trivial codes (no checks) behave sensibly without special-case guards."""
-    code = CSSCode.get_trivial_code(3)
-    e = np.array([1, 0, 0], dtype=np.int8)
-    zero = np.zeros(3, dtype=np.int8)
-    assert code.get_x_syndrome(e).size == 0
-    assert code.get_z_syndrome(e).size == 0
-    assert code.check_if_x_stabilizer(zero)
-    assert not code.check_if_x_stabilizer(e)
-    assert code.check_if_z_stabilizer(zero)
-    assert not code.check_if_z_stabilizer(e)
-    assert code.stabilizer_eq_x_error(e, e)
-    assert not code.stabilizer_eq_x_error(e, zero)
-    assert code.check_if_logical_x_error(e)
-    assert code.check_if_logical_z_error(e)
